Loading best and castle files
best_armies.h
A header containing only those armies which won 70% or more of their wars in the first Riddler.

castle.c
A C program to compute the won-lost records of all entries in the file.

solve_best.c
This is the same as solve_castle.c, but using a different header file of entries. Yes, ideally this program would just read a .csv as input.

castle.pl
Quickly written, but slower to run, version of castle.c. This was actually the first program I wrote for this puzzle.